requirements.txt
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.5
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
blinker==1.7.0
boilerpy3==1.0.7
cachetools==5.3.2
cattrs==23.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-haystack==0.8.1
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.14
distro==1.8.0
docopt==0.6.2
Events==0.5
farm-haystack==1.22.1
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.1
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.17.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflect==7.0.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kubernetes==28.1.0
langdetect==1.0.9
lazy-imports==0.3.1
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.7
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pdf2image==1.16.3
Pillow==10.1.0
platformdirs==4.0.0
portalocker==2.8.2
posthog==3.0.2
prompthub-py==4.0.0
protobuf==4.25.1
psutil==5.9.6
pulsar-client==3.3.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
PyPika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-frontmatter==1.0.1
python-magic-bin==0.4.14
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.7.0
qdrant-haystack==1.0.11
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.16.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.50
SQLAlchemy-Utils==0.41.1
sseclient-py==1.8.0
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tika==2.6.0
tiktoken==0.5.1
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformers==4.34.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
url-normalize==1.4.3
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
